# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IAuthor": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IReview": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"bookId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"rating",
					"createdAt",
					"userId",
					"bookId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.name-or-username-or-profilePicture_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"profilePicture"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ReviewUser": {
				"$ref": "#/components/schemas/Pick_IUser.name-or-username-or-profilePicture_"
			},
			"IReviewWithUser": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IReview"
					},
					{
						"properties": {
							"user": {
								"$ref": "#/components/schemas/ReviewUser"
							}
						},
						"required": [
							"user"
						],
						"type": "object"
					}
				]
			},
			"Pick_IBook.Exclude_keyofIBook.authors-or-genres__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"coverPicture": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"isbn": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"format": {
						"type": "string",
						"enum": [
							"PAPERBACK",
							"HARDCOVER",
							"EBOOK"
						]
					},
					"language": {
						"type": "string"
					},
					"numOfPages": {
						"type": "number",
						"format": "double"
					},
					"pdfLink": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"country",
					"title",
					"coverPicture",
					"rating",
					"isbn",
					"description",
					"publishDate",
					"format",
					"language",
					"numOfPages",
					"pdfLink"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IBook.authors-or-genres_": {
				"$ref": "#/components/schemas/Pick_IBook.Exclude_keyofIBook.authors-or-genres__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IBookWithoutAuthorsAndGenres": {
				"$ref": "#/components/schemas/Omit_IBook.authors-or-genres_"
			},
			"IReviewWithBook": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IReview"
					},
					{
						"properties": {
							"book": {
								"$ref": "#/components/schemas/IBookWithoutAuthorsAndGenres"
							}
						},
						"required": [
							"book"
						],
						"type": "object"
					}
				]
			},
			"IReviewWithUserAndBook": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IReviewWithUser"
					},
					{
						"$ref": "#/components/schemas/IReviewWithBook"
					}
				]
			},
			"IBookshelf": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"privacy": {
						"type": "string",
						"enum": [
							"PRIVATE",
							"PUBLIC"
						]
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"books": {
						"items": {
							"$ref": "#/components/schemas/IBookWithoutAuthorsAndGenres"
						},
						"type": "array"
					},
					"_count": {
						"properties": {
							"books": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"books"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"createdAt",
					"privacy",
					"userId",
					"books",
					"_count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGenre": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBook": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"isbn": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"format": {
						"type": "string",
						"enum": [
							"PAPERBACK",
							"HARDCOVER",
							"EBOOK"
						]
					},
					"language": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"numOfPages": {
						"type": "number",
						"format": "double"
					},
					"pdfLink": {
						"type": "string",
						"nullable": true
					},
					"coverPicture": {
						"type": "string",
						"nullable": true
					},
					"authors": {
						"items": {
							"$ref": "#/components/schemas/IAuthor"
						},
						"type": "array"
					},
					"genres": {
						"items": {
							"$ref": "#/components/schemas/IGenre"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"isbn",
					"description",
					"publishDate",
					"format",
					"language",
					"country",
					"rating",
					"numOfPages",
					"pdfLink",
					"coverPicture",
					"authors",
					"genres"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BookshelfUser": {
				"$ref": "#/components/schemas/Pick_IUser.name-or-username-or-profilePicture_"
			},
			"IBookshelfWithUser": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IBookshelf"
					},
					{
						"properties": {
							"user": {
								"$ref": "#/components/schemas/BookshelfUser"
							}
						},
						"required": [
							"user"
						],
						"type": "object"
					}
				]
			},
			"OptionalBookshelf": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IBookshelf"
					}
				],
				"nullable": true
			},
			"Privacy": {
				"enum": [
					"PUBLIC",
					"PRIVATE"
				],
				"type": "string"
			},
			"CreateBookshelfDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"privacy": {
						"$ref": "#/components/schemas/Privacy"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"description",
					"privacy",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBookshelfDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"privacy": {
						"$ref": "#/components/schemas/Privacy"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IBookshelf.Exclude_keyofIBookshelf.books-or-_count__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"privacy": {
						"type": "string",
						"enum": [
							"PRIVATE",
							"PUBLIC"
						]
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"createdAt",
					"privacy",
					"userId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IBookshelf.books-or-_count_": {
				"$ref": "#/components/schemas/Pick_IBookshelf.Exclude_keyofIBookshelf.books-or-_count__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IBookshelfWithoutBooks": {
				"$ref": "#/components/schemas/Omit_IBookshelf.books-or-_count_"
			},
			"OptionalBook": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IBook"
					}
				],
				"nullable": true
			},
			"Format": {
				"enum": [
					"PAPERBACK",
					"HARDCOVER",
					"EBOOK"
				],
				"type": "string"
			},
			"CreateBookDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"isbn": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"format": {
						"$ref": "#/components/schemas/Format"
					},
					"country": {
						"type": "string"
					},
					"numOfPages": {
						"type": "number",
						"format": "double"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"pdfLink": {
						"type": "string"
					},
					"coverPicture": {
						"type": "string"
					},
					"authors": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"genres": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"isbn",
					"description",
					"language",
					"format",
					"country",
					"numOfPages",
					"publishDate",
					"pdfLink",
					"coverPicture",
					"authors",
					"genres"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthorDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenreDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBookDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"isbn": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"format": {
						"$ref": "#/components/schemas/Format"
					},
					"country": {
						"type": "string"
					},
					"numOfPages": {
						"type": "number",
						"format": "double"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"pdfLink": {
						"type": "string"
					},
					"coverPicture": {
						"type": "string"
					},
					"authors": {
						"items": {
							"$ref": "#/components/schemas/AuthorDto"
						},
						"type": "array"
					},
					"genres": {
						"items": {
							"$ref": "#/components/schemas/GenreDto"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"isbn",
					"description",
					"language",
					"format",
					"country",
					"numOfPages",
					"publishDate",
					"pdfLink",
					"coverPicture",
					"authors",
					"genres"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "grad-proj-core",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/{userId}/bookshelves": {
			"get": {
				"operationId": "GetBookshelvesByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IBookshelf"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/bookshelves": {
			"get": {
				"operationId": "GetAllBookshelves",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IBookshelfWithUser"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "term",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBookshelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookshelf"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookshelfDto"
							}
						}
					}
				}
			}
		},
		"/bookshelves/{bookshelfId}": {
			"get": {
				"operationId": "GetBookshelfById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/OptionalBookshelf"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookshelfId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBookshelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookshelf"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookshelfId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookshelfDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBookshelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookshelfWithoutBooks"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookshelfId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bookshelves/add-books/{bookshelfId}": {
			"patch": {
				"operationId": "AddBookToBookshelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookshelf"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookshelfId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"bookIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"bookIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/bookshelves/remove-books/{bookshelfId}": {
			"patch": {
				"operationId": "RemoveBooksFromBookshelf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookshelf"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Bookshelves"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookshelfId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"bookIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"bookIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/books": {
			"get": {
				"operationId": "GetAllBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IBook"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "term",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookWithoutAuthorsAndGenres"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookDto"
							}
						}
					}
				}
			}
		},
		"/books/{bookId}": {
			"get": {
				"operationId": "GetBookById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/OptionalBook"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBookById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookWithoutAuthorsAndGenres"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBookById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IBookWithoutAuthorsAndGenres"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/books/{bookId}/reviews": {
			"get": {
				"operationId": "GetReviewsByBookId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IReviewWithUserAndBook"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/books/{bookId}/authors": {
			"get": {
				"operationId": "GetAuthorsByBookId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IReviewWithUserAndBook"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Books"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}